# Reduce size of ZSTD_DCtx by tapping dst during decompression to reduce literal buffer