Stores the literal buffer at the end of the destination buffer to reduce memory requirements in ZSTD_DCtx.

----

Still fixing a couple edge cases before it's ready for landing, after fixing the issue that was causing the fallback loop to be used over-aggressively, I did some performance testing and having a fallback loop and alternate storage offset conditions seems to be unnecessary.  I changed to this approach, which is far simpler and seems to perform adequately.

With changes:

binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P00 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 32372.4 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  1235.0 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2059.1 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2746.5 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 14646.9 MB/s  ( 1000000) 



Without changes:

binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P00 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 32371.2 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  1256.8 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2032.0 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2618.5 MB/s  ( 1000000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B1000000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug  1 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 14662.3 MB/s  ( 1000000)