@binhdvo 2021-08-12
Updated fixing the issues with incorrect decompression and failing fullbench, still following up on some fuzzer-related issues and gathering thorough performance numbers.