@binhdvo 2021-08-12
Performance takes a hit on small files at high compressibility, it may be necessary to return the second cleanup loop to address this:



with changes 100k

*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   : 47585.9 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B100000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  1164.1 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B100000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  1992.2 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B100000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  3162.2 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B100000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  3882.7 MB/s  (  100000) 



without changes 100k

binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P0 -B100000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   : 49230.5 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B100000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  1152.2 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B100000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  1950.7 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B100000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  2907.3 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B100000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  6571.4 MB/s  (  100000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B100000 



with changes 10k

*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   : 34960.1 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B10000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1209.2 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B10000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1537.9 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B10000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  2787.5 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B10000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  3704.3 MB/s  (   10000) 



without changes 10k

*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   : 33984.8 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B10000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1179.6 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B10000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1545.6 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B10000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  2924.4 MB/s  (   10000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B10000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  5829.1 MB/s  (   10000) 



with changes 1k

*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  4537.6 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B1000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   526.9 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B1000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   528.8 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B1000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  1578.0 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B1000  
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 11 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  1775.3 MB/s  (    1000) 



without changes 1k

*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  4310.9 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P10 -B1000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   507.0 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P50 -B1000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   526.9 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P90 -B1000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  2265.4 MB/s  (    1000) 
binhvo@binhvo-mbp zstd % ./tests/fullbench -b2 -P100 -B1000
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 10 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  2287.2 MB/s  (    1000)